SystemUI: remember deleted custom tiles
We previously had no mechanism of knowing which tiles the user had
manually removed, leading to the tiles returning when the publishing
application decided it was time to update.

Now, when the user removes a custom tile; set a flag and hide it in the
tile list. The removed tiles will be enumerated in the Add tile detail
view if the user decides they want it back.

Ticket: CYNGNOS-2530

Change-Id: Ia4fe74ef580af7139279a2fb78e240e872f8dbb0