Keyguard: Allow swiping up when trust managed
When the lock screen is being trust managed, the center keyguard
affordance view becomes clickable and consumes touch events, which
includes the user swiping up from the center icon to unlock.  This
patch checks if the lock screen is being trust managed and if so,
intercepts the touch events while retaining the clickability of the
center lock icon.

TICKET: CYNGNOS-2687

Change-Id: I54dd9bfdded3dbd2bf90ed1fc49dd55e5b3a7214