WindowAnimator: Check KEYGUARD_PANEL when deciding to show/hide window
By default, if blur is enabled we don't hide the window behind since we want the user
to see it. This is not true for when a KEYGUARD_PANEL is active since we can't see
anything behind it anyway. Only apply blur logic if KEYGUARD_PANEL is not added.

CYNGNOS-2980

Change-Id: Icecae14455d6e68665bb688e3f6dc4ea9d74e88c